requirements.txt
bottle
qrcode
Pillow
pypi
requests
bcrypt
pymongo
